refactor(archtest): FIXTURE-CELLID-TYPED-BUILDER-01 + METADATATEST-IMPORT-SCOPE-01 迁 module-path-agnostic + 缩 baseline [#1639]
移两条 invariant 的规则体 + 全部 scanner helper 到非 test
fixture_cellid_typed_builder.go（metadata/metadatatest pkg 路径 + carve-out key
+ A2 in-func panicreg/errcode 路径派生自 PlatformModulePath）。_test.go 瘦身为
A1/import-scope dogfood + 完整保留 A2/A5 body-shape、A3 negative RED fixture、
A4 ADR-consistency 自检（anti-vacuity 不弱化）。register=none（kernel/metadata
结构内化，外部 vacuous）。baseline 删该行。
